refactor(toast): move `ToastProvider` outside of internals
