+++ packages/react/src/aksara-provider/AksaraProvider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { IdProvider } from '@radix-ui/react-id';
 import { ThemeProvider } from 'styled-components';
-import { ToastProvider } from '../components/toast/internals';
+import { ToastProvider } from '../components/toast';
 
 import { theme as defaultTheme, Theme } from '../theme';
 import injectGlobalStyles from './injectGlobalStyles';
