+++ packages/react/src/components/toast/index.ts
@@ -1,2 +1,3 @@
 export { default as useToast } from './useToast';
+export { default as ToastProvider } from './ToastProvider';
 export * from './types';
