+++ packages/react/src/components/toast/internals/index.ts
@@ -1,2 +1,3 @@
 export { default as ToastContext } from './ToastContext';
-export { default as ToastProvider } from './ToastProvider';
+export { default as ToastContainer } from './ToastContainer';
+export * from './constants';
